Revert 587a914bfe and b2535c2eae.
Something about the `cargo update` made tests start failing on Mac in a
non-deterministic way. Though `cargo test -- --test-threads=1` ran ok,
so something to do with test parallelism. Maybe a problem with
`NamedTempFile`?

So this commit undoes the `cargo update`. It also undoes the switch from
`regex` to `regex_lite`, which was necessary to undo the `cargo update`.
I'll redo the `regex` to `regex_lite` change shortly.