Avoid allocating on every input line.
By using `BufRead::read_line` instead of `BufRead::lines`. Now
allocations are only required (a) for the first occurrence of a
particular line, and (b) for lines that are modified by `-e`.

This requires some changes to handle cases where the line is a `&str` vs
a `String` (e.g. after modification due to `-e`).

This roughly doubles the speed of `counts` on files where every line is
the same, and leaves it unchanged on files where every line is
different. Typical cases will be between those two extremes.

This idea came from @Shnatsel via
https://github.com/nnethercote/perf-book/issues/68.